Moved files out of i18n/namespaces folder into main folder.
