+++ app/i18n/en/contentItems.js
