+++ app/i18n/en/editor.js
