+++ app/i18n/en/flash.js
